Create NegativeWeightCycleDetection.java
You are given 2 integers N and M , N is the number of vertices, M is the number of edges. You'll also be given ai, bi and wi where ai and bi represents an edge from a vertex ai to a vertex bi and wi respresents the weight of that edge.

Task is to find if it contains a negative weight cycle or not.


Print 1 if it contains a negative weight cycle else print 0.